requirements.txt
openai
streamlit
faiss-cpu
sentence-transformers
pandas
openai-whisper
yt-dlp
requests